requirements.txt
# Web framework
flask
flask-cors

# UI
gradio>=4.0.0

# AI / ML
sentence-transformers>=2.6.0
transformers>=4.37.0
torch
torchvision

# Image processing
pillow

# Data
numpy

# HTTP
requests
